+++ src/main/java/kitchenpos/products/application/ProductService.java
@@ -1,6 +1,6 @@
 package kitchenpos.products.application;
 
-import kitchenpos.common.domain.PurgomalumClient;
+import kitchenpos.products.domain.PurgomalumClient;
 import kitchenpos.menus.domain.Menu;
 import kitchenpos.menus.domain.MenuProduct;
 import kitchenpos.menus.domain.MenuRepository;
