+++ src/main/java/kitchenpos/products/infra/DefaultPurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.common.infra;
+package kitchenpos.products.infra;
 
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Component;
@@ -7,7 +7,7 @@ import org.springframework.web.util.UriComponentsBuilder;
 
 import java.net.URI;
 
-import kitchenpos.common.domain.PurgomalumClient;
+import kitchenpos.products.domain.PurgomalumClient;
 
 @Component
 public class DefaultPurgomalumClient implements PurgomalumClient {
