+++ src/test/java/kitchenpos/application/FakePurgomalumClient.java
@@ -1,6 +1,6 @@
 package kitchenpos.application;
 
-import kitchenpos.common.domain.PurgomalumClient;
+import kitchenpos.products.domain.PurgomalumClient;
 
 import java.util.Arrays;
 import java.util.List;
